drivers: disk: add fakefat disk driver
Fakefat disk emulates a disk formatted with the FAT file system.
The driver supports FAT16 and FAT32 emulation and multiple instances.
A backend can register at least 15 emulated files on a disk instance.
The file structure contain read and write callbacks that are called
when the filesystem driver accesses the volume. A backend can have
multiple files with read callbacks, but usually only one or zero with
write callbacks.

Fakefat disk can be used for testing or to map services such as
firmware updates to a file on the emulated disk.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>