net: conn_mgr: Implement online connectivity status check
Add a way to check if we have an Internet connectivity before
emitting L4_CONNECTED event. We support two strategies for this,
either an ICMP Echo-Request (ping) to a user defined host, or
a HTTP GET request to a user defined URL. Both IPv4 and IPv6 checks
are supported. If the user defined host is not an IP, then DNS
is used to resolve the host address.

Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>