include/zephyr: Add a new macro __malloc_like_with_free()
GCC 11 introduced[1] a new function attribute malloc, which, as the GCC
document says, indicates that a function is ‘malloc’-like.

With this attribute, GCC can find a silly use-after-free. Given that:

    #include <zephyr/kernel.h>

    struct Foo { int a; };

    int main(void)
    {
        struct Foo *p = k_malloc(sizeof(struct Foo));
        if (p) {
            k_free(p);
            p->a = 1;
        }
        return 0;
    }

GCC can now warn you:

    use-after-free/src/main.c: In function 'main':
    use-after-free/src/main.c:10:22: warning: pointer 'p' used after
    'k_free' [-Wuse-after-free]
       10 |                 p->a = 1;
          |                 ~~~~~^~~
    use-after-free/src/main.c:9:17: note: call to 'k_free' here
        9 |                 k_free(p);
          |                 ^~~~~~~~~

The macro name, __malloc_like_with_free, is chosen to align with picolibc
since there are some usage of __malloc_like in zephyr modules.

[1]:https://gcc.gnu.org/bugzilla/show_bug.cgi?id=94527

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>