dts: zynq7000: add the clock control driver to the Zynq-7000 DT
Add the node for the clock control driver below the SLCR syscon
node in the Zynq-7000 DT.

At the time being, the modified DT also contains two nodes that
specify an EMIO clock configuration for the two integrated
Ethernet controller, in order to demonstrate what a clock
specification that cannot be derived from one of the Zynq's
three main PLLs (ARM/DDR/IO PLL) looks like. EMIO clocks are
provided from the programmable logic part of the chip and just
provide a fixed frequency (to which divisors can in turn be
applied) that cannot be read out of the running system in any
other way.

Signed-off-by: Immo Birnbaum <mail@birnbaum.immo>