scripts: pylib: twister: Fix the unrecognized section test
Current implementation of the unrecognized section test
does not work at all. Here we move its handling into the
runner.py's process() method so it's not scattered across
three different files, which could be one of the reasons
for its failure before.

Additionally, a blackbox test was added to prove correctness.

Signed-off-by: Lukasz Mrugala <lukaszx.mrugala@intel.com>