dts: arm: renesas: ra: Improved device tree organization
For commonality, change ra-cm4-common.dtsi to only basic definitions
such as node names and registers.

The configuration for each series (series in soc.yaml, which corresponds to
a group rather than series in the Renesas product description) is somewhat
common, so we will create a common definition as r7fa4m1xxxxxx.dtsi.
The definition of peripheral interrupts differs depending on the series,
so we define them here.

Each SoC model's definition file (r7fa4m1ab3cfm.dtsi) reflects the
available pin settings.

Also, in the original configuration of cm4, a non-existent irq9 was
set for ioport2, but this was unnecessary so I deleted it.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>