tests: drivers: clock_mgmt: add clock management hardware test
Add clock management hardware test. This test applies a series of clock
states for a given consumer, and verifies that each state produces the
expected rate. This test is intended to verify that each clock node
driver within an SOC implementation works as expected.

Boards should define their test overlays for this test to exercise as
much of the clock tree as possible, and insure some clock-output nodes
are used as well, to test their `clock_set_rate` and `clock_round_rate`
implementations.

Initial support is added for the `lpcxpresso55s69/lpc55s69/cpu0` target,
as this is the only hardware supporting clock management.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>