drivers: misc: Add support for WIZ550io module
WIZ550io is a W5500 breakout board with an extra PIC12F519 MCU.
The PIC12F519 MCU initialises the W5500 with a unique MAC address
after a GPIO reset. The PIC12F519 requires a 150ms delay after GPIO
reset to configure the W5500. There is an additional "ready" GPIO
signal that can be used to reduce that delay.

Signed-off-by: Grant Ramsay <gramsay@enphaseenergy.com>