dts: arm: st: add npgios to all stm32 gpio controllers
For almost all STM32 GPIO controllers, the number of supported GPIO
pins managed by a single controller is 16 (with some exceptions for
fewer). However, the default for ngpios in the device tree bindings
for gpio-controllers is 32; leading to inaccuracies in handling GPIO
for these controllers, such as presenting too many GPIOs in the GPIO
shell. This patch adds the ngpios property for all currently supported
STM32 GPIO controllers.

Signed-off-by: Michael R Rosen <mrrosen@alumni.cmu.edu>