timer: start directly in init.c
Do not use SYS_INIT, directly call init_sys_clock_driver() during the
init process.

The system timer driver is not a traditional driver and its dependencies
are fixed. It is needed at a fixes location during the boot process and
is not something you want to change or tamper with. Any change to where
the timer is started might have negative consquences and non-functioning
system. Thus, call this "driver" and initialize it at a fixed location
and avoid the SYS_INIT use.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>