pm: Add support to invoke Turn on and off actions
The current implementation of system managed device power
management only invokes the suspend and resume actions on
the device drivers.
This change allows the user to specify a property per
power state called zephyr,pm-turn-on-off-action. This will
pass turn on and off actions to the device driver for that
power state.
This allows device drivers to take additional power saving
steps for certain low power states.

Signed-off-by: Mahesh Mahadevan <mahesh.mahadevan@nxp.com>