cmake: version.cmake: remove tweak from {PROJECT}VERSION
The `KERNELVERSION` and `APPVERSION` integers in the generated version
files are documented in `kernel_version.h` as having the last byte
reserved, not set to the tweak value.

The last byte in this integer is by necessity only 8 bits, while there
is no such restriction on the version tweak value.

Signed-off-by: Jordan Yates <jordan@embeint.com>